move CustomMetaTags and OpenGraphScraperOptions to types and convert utils to start using some types
